+++ R/RcppExports.R
@@ -53,7 +53,7 @@ write_sas_ <- function(data, path) {
     invisible(.Call(`_haven_write_sas_`, data, path))
 }
 
-write_xpt_ <- function(data, path, version) {
-    invisible(.Call(`_haven_write_xpt_`, data, path, version))
+write_xpt_ <- function(data, path, version, name) {
+    invisible(.Call(`_haven_write_xpt_`, data, path, version, name))
 }